pyproject.toml
[tool.poetry]
name = "picmory-be"
version = "0.1.0"
description = ""
authors = ["Ksanbal <dev.ksanbal@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.9"
django = "4.2.13"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.2.2"
pylint-django = "^2.5.5"
yapf = "^0.40.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
